Job specific mount paths
So persistent volume mounts for different jobs on the same host don't
clash, have job specific paths for the mounts.

Also this tries to make the mounts consistent between what is done and
what is reported via environment variables. This included needing to
make a symlink for private per job buffers so the per compute host
location is the same on all compute nodes, yet still private to each
node.