Add mutex to volume attachment updates
With lots of calls to add/remove attachments to persistent volumes we
hit races where we say updates clash and (correctly) get rejected by
etcd. This should stop that, although really a restructure of the way we
store data might avoid this problem in a better way, this should get us
up and running for now.

The state updates don't need this additional lock, because of the way
the updates flow between process, and the race detection we get from how
we already use etcd.