Hold lock while doing allocations
Try to hold a lock when querying the state of the pool, choosing some
bricks, and assigning those bricks to a specific volume.

We could do some retry loops to avoid locks, but there will be
contention on these resources any way we look at this problem.

To hide this complexity behind the correct interface, the pool registry
was made responsible for selecting the bricks.