Increase timeouts for mount and ansible
The mount timeout was only 1 min, even though the log message said
it was 5 mins. Fixes: #122.

It seems 5 mins isn't long enough for some ansible runs, even when
you increase the ansible forking to equal the number of DAC nodes.
For now try bump this to 10 mins, but leave bug #121 for the config
idea open for now.