Ensure MDTs index is sequential. (#128)
Lustre was getting confused about missing MDTs when we had buffer sizes
that mean we started to limit the number of MDTs. You would see
disconnection errors in the logs.

At the same time, do a better job of limiting the max number of MDTs
when you have a large number of NVMe disks spread over a smaller number
of hosts.