Add option to override MGS host (#132)
Added the config:
DAC_MGS_HOST

By default it is localhost, which is existing behaviour where
MGS always lives on the primary brick host.

We found operational issues of when to clean out the MGS after
creating lots of filesystems. Eventually the MGS will fill up.
As a work around, we allow operators to specify a custom location
(host and device) for the MGS. This means you could easily rotate
a single MGS (via change of config followed by reboot of dacd)
that ensures the correct thing happens.